refactor using viewModel
